Moved auth gens from RequestTest.
