+++ consumer-data-au-api-types/tests/Web/ConsumerData/Au/Api/Types/Auth/Gens.hs
@@ -6,12 +6,13 @@ module Web.ConsumerData.Au.Api.Types.Auth.Gens where
 import qualified Data.Dependent.Map as DM
 import           Data.Dependent.Sum (DSum ((:=>)))
 
-import           Hedgehog       (MonadGen)
-import qualified Hedgehog.Gen   as Gen
-import qualified Hedgehog.Range as Range
+import           Control.Monad.IO.Class (MonadIO, liftIO)
+import qualified Crypto.JOSE.JWK        as JWK
+import           Hedgehog               (MonadGen)
+import qualified Hedgehog.Gen           as Gen
+import qualified Hedgehog.Range         as Range
 
-import Web.ConsumerData.Au.Api.Types.Auth.AuthorisationRequest
-    (Claims (Claims))
+import Web.ConsumerData.Au.Api.Types.Auth.AuthorisationRequest (Claims (Claims))
 import Web.ConsumerData.Au.Api.Types.Auth.Common.Common
     (Acr (Acr), Claim (Claim), TokenSubject (..))
 import Web.ConsumerData.Au.Api.Types.Auth.Common.IdToken
@@ -43,3 +44,22 @@ genClaims ::
   => n Claims
 genClaims =
   Claims Nothing <$> genIdTokenClaims
+
+genJWK ::
+  ( MonadGen n
+  , MonadIO n
+  )
+  => n JWK.JWK
+genJWK =
+  genKeyMaterial >>= liftIO . JWK.genJWK
+
+-- | Valid key material dictated by allowed signing algorithms (see 'signingAlg') and the
+-- <https://github.com/frasertweedale/hs-jose/blob/18865d7af9d3b16d737f38579643399cf4facc1b/src/Crypto/JOSE/JWA/JWK.hs#L610 JWK module in @jose@>
+genKeyMaterial ::
+  MonadGen n
+  => n JWK.KeyMaterialGenParam
+genKeyMaterial =
+  Gen.choice
+    [ pure (JWK.ECGenParam JWK.P_256)
+    , JWK.RSAGenParam <$> Gen.int (Range.linear (2048 `div` 8) (4096 `div` 8))
+    ]
